pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "idea-gen"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "duckduckgo-search>=6.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
